+++ examples/resources/nxos_feature_pim/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_feature_pim.example "sys/fm/pim"